==> R/00_all.qmd <==
---
title: "99_render_to_html"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load packages

```{r}
library(here)
```

### Render all documents to html in results folder

```{r}
rmarkdown::render(here("R/01_load.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/02_clean.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/03_augment.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/04_describe.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/05_analysis_heatmap.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/06_analysis_scatterplot.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/07_analysis_stats.qmd"), 
                  output_dir = here("results"))
rmarkdown::render(here("R/08_analysis_kegg.qmd"), 
                  output_dir = here("results"))
system(paste("quarto render", here("doc/presentation.qmd")))
```
