==> R/01_load.qmd <==
---
title: "01_load"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load libraries

```{r}
library(tidyverse)
library(readr)
library(here)
```

### Load data and save for later use

The data is loaded as tibbles (tsv files) from four different txt files,
each containing a column with gene symbols and a column with expression
levels. For the genes where the expression level is empty, NA or N/A, it
is saved as "na".

```{r}
#First load the raw data taking care of missing values 
SKOV3_MOCK <- read_tsv(file = here("data/_raw/GSM8331622_SKOV_MOCK.txt"))
SKOV3_KO <- read_tsv(file = here("data/_raw/GSM8331624_SKOV_KO.txt"))
OVCAR3_MOCK <- read_tsv(file = here("data/_raw/GSM8331623_OVCAR3_MOCK.txt")) 
OVCAR3_KO <- read_tsv(file = here("data/_raw/GSM8331625_OVCAR_KO.txt")) 
```

### Removing duplicates

If a gene occurs more than once, all duplicates are removed and the
specific gene does not occur in the analysis. There is not sufficient
information about why duplicates occur, this is why we have chosen to
remove all cases.

```{r}
OVCAR3_KO <- OVCAR3_KO |>
  group_by(`Gene Symbol`) |> 
  filter(n() == 1) |> 
  ungroup()

OVCAR3_MOCK <- OVCAR3_MOCK |>
  group_by(`Gene Symbol`) |>
  filter(n() == 1) |>
  ungroup()

SKOV3_KO <- SKOV3_KO |>
  group_by(`Gene Symbol`) |>
  filter(n() == 1) |>
  ungroup()

SKOV3_MOCK <- SKOV3_MOCK |>
  group_by(`Gene Symbol`) |>
  filter(n() == 1) |>
  ungroup()
```

### Joining data

The four tibbles are joined into one tibble with rows corresponding to
the unique genes and columns corresponding to each cell line (OVCAR3 and
SKOV3) at baseline (\_MOCK) and after knock-out of gene HERV-K (\_KO):

OVCAR3_MOCK, OVCAR3_KO, SKOV3_MOCK, and SKOV3_KO.

```{r}
data <- left_join(OVCAR3_MOCK,
                  OVCAR3_KO, 
                  by = join_by(`Gene Symbol`)) |> 
  left_join(SKOV3_MOCK, 
            by = join_by(`Gene Symbol`)) |> 
  left_join(SKOV3_KO, 
            by = join_by(`Gene Symbol`)) |> 
  rename(gene = `Gene Symbol`)
```

### Saving in data folder

```{r}
write_tsv(data, here("data", "01_dat_load.tsv"))
```
