==> R/02_clean.qmd <==
---
title: "02_clean"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load packages

```{r}
library(tidyverse)
library(readr)
library(here)
```

### Load data

```{r}
data <- read_tsv(file = here("data/01_dat_load.tsv"))
```

### Clean up data

The data is cleaned by removing all rows (genes) that has an expression
level of 0 in all four cases. The columns are renamed to ensure
alignment in names so that the cell lines are called OVCAR3 and SKOV3
respectively, (OVCAR3_MOCK, OVCAR3_KO, SKOV3_MOCK, SKOV3_KO).

```{r}
clean_data <- data |> 
  filter(rowSums(across(where(is.numeric))) > 0) |> 
  rename(OVCAR3_KO = "OVCAR_KO",
         SKOV3_KO = "SKOV_KO",
         SKOV3_MOCK = "SKOV_MOCK")
```

### Save data for later use

```{r}
write_tsv(clean_data, here("data", "02_dat_clean.tsv"))
```
