==> R/04_describe.qmd <==
---
title: "04_describe"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load packages

```{r}
library(tidyverse)
library(readr)
library(here)
```

### Load data

```{r}
data <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

### Basic information

```{r}
ncol(data)
nrow(data)
```

There are 28.040 observations (genes).

There are 9 variables, these include: MOCK, KO, log2FC and DEG
(up/downregulated) for each cell line and a column with gene symbols.

### Number of up-regulated and down-regulated genes

The number of up- and down-regulated genes are counted.

```{r}
data |> 
  count(DEG_OVCAR3)

data |> 
  count(DEG_SKOV3)
```

After HERV-K KO in the OVCAR3 cell line, there are 766 down-regulated
genes, 4.212 up-regulated genes and 23.062 neutral genes.

After HERV-K KO in the SKOV3 cell line, there are 920 down-regulated
genes, 270 up-regulated genes and 26.850 neutral genes.

A barplot is made to illustrate the number of differentially expressed
genes (DEG) for each cell line. They are divided into down-regulated,
up-regulated and neutral genes.

```{r}
barplot <- data |> 
  pivot_longer(
    cols = c(DEG_OVCAR3, 
             DEG_SKOV3), 
    names_to = "cell_line", 
    values_to = "expression_level") |> 
  mutate(cell_line = case_when(cell_line == "DEG_OVCAR3" ~ "OVCAR3",
                               cell_line == "DEG_SKOV3" ~ "SKOV3")) |>
  select(gene, cell_line, expression_level) |> 
  count(expression_level, 
        cell_line) |> 
  ggplot(mapping = aes(x = expression_level, 
                       y = n, 
                       fill = cell_line)) +
  geom_bar(stat = "identity", 
           position = "dodge") +
  geom_text(mapping = aes(label = n), 
            position = position_dodge(width = 0.9),
            vjust = -0.2) +
  theme_minimal() + 
  theme(plot.title = element_text(size = 24),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 18)) +
  labs(x = "",
       y = "Count",
       title = "Distribution of DEGs in the two cell lines OVCAR3 \nand SKOV3",
       fill = "Cell line") + 
  scale_x_discrete(limits = c("Downregulated", "Upregulated", "Neutral"), 
                   labels = c("Downregulated" = "Down-regulated", "Upregulated" = "Up-regulated"))

```

The barplot is saved in a png-file in the results folder::

```{r}
ggsave(filename = here("results/04_gene_counts.png"),
       plot = barplot,
       width = 10, 
       height = 6)
```

### Number of overlapping genes between the cell lines

Counting the number of genes that are up-regulated or down-regulated in
both cell lines.

```{r}
#upregulated genes for both cell lines
data |> 
  filter(DEG_OVCAR3 == "Upregulated" &
         DEG_SKOV3 == "Upregulated") |> 
  print() |> 
  count()

#downregulated genes for both cell lines
data |> 
  filter(DEG_OVCAR3 == "Downregulated" &
         DEG_SKOV3 == "Downregulated") |>
  print() |>
  count()
```

For the up-regulated genes there are 59 genes that are overlapping in
both cell lines, CRYBB2 and ANKRD62 both associated with cancer.

In regards to the down-regulated genes there are 15 genes shared between
the two cell lines, RBAKDN and CDH2 both associated with cell division
and adhesion.
