==> R/05_analysis_heatmap.qmd <==
---
title: "05_analysis_heatmap"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load packages

```{r}
library(tidyverse)
library(here)
library(readr)
```

### Load data

```{r}
data <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

### Heatmap of randomly selected genes

A heatmap of 70 randomly sampled genes is created. The heatmap shows the
70 genes colored by log2 fold change for each cell line. Red color
indicates up-regulated genes and blue indicates down-regulated genes.

```{r}
#70 are sampled randomly
heatmap_random <- data |>
  sample_n(70) |> 
  pivot_longer(
    cols = c(log2FC_OVCAR3, log2FC_SKOV3), 
    names_to = "cell_line", 
    values_to = "log2_expression") |>
  ggplot(mapping = aes(x = gene,
                       y = cell_line,
                       fill = log2_expression)) +
  geom_tile() +
  scale_fill_gradient2(midpoint = 0,
                       low = "blue",
                       mid = "white",
                       high = "red") +
  labs(x = "Gene",
       y = "Cell line",
       title = "Gene expression differences from baseline to HERV-K KO in \nSKOV3 and OVCAR3 cells",
       fill = "Log2FC") +
  theme_classic() +
  theme(plot.title = element_text(size = 24),
        axis.title.x = element_text(size = 20),
        axis.text.x = element_text(size = 6,
                                   angle = 45, 
                                   hjust = 1),
        axis.title.y = element_text(size = 20),
        axis.text.y = element_text(size = 16),
        legend.text = element_text(size = 14),
        legend.title = element_text(size = 18)) +
  scale_y_discrete(labels = c("log2FC_SKOV3" = "SKOV3", "log2FC_OVCAR3" = "OVCAR3"))

heatmap_random
```

Saving the heatmap in a png-file in the results folder:

```{r}
ggsave(filename = here("results/05_heatmap_random.png"),
       plot = heatmap_random,
       width = 12, 
       height = 8)
```

### Heatmap of overlapping up- and down-regulated DEGs in the two cell lines

```{r}
heatmap_overlap <- data |>
  pivot_longer(
    cols = c(log2FC_OVCAR3, log2FC_SKOV3), 
    names_to = "cell_line", 
    values_to = "log2_expression") |>
  filter(DEG_OVCAR3 == "Upregulated" &
           DEG_SKOV3 == "Upregulated" |
           DEG_OVCAR3 == "Downregulated" &
           DEG_SKOV3 == "Downregulated") |>
  ggplot(mapping = aes(x = gene,
                       y = cell_line,
                       fill = log2_expression)) +
  geom_tile() +
  scale_fill_gradient2(midpoint = 0,
                       low = "blue",
                       mid = "white",
                       high = "red") +
  labs(x = "Gene",
       y = "Cell line",
       title = "Gene expression differences from baseline to HERV-K KO in \nSKOV3 and OVCAR3 cells for overlapping DEGs",
       fill = "Log2FC") +
  theme_classic() +
  theme(plot.title = element_text(size = 24),
        axis.title.x = element_text(size = 20),
        axis.text.x = element_text(size = 6,
                                   angle = 45, 
                                   hjust = 1),
        axis.title.y = element_text(size = 20),
        axis.text.y = element_text(size = 16),
        legend.text = element_text(size = 14),
        legend.title = element_text(size = 18)) +
  scale_y_discrete(labels = c("log2FC_SKOV3" = "SKOV3", "log2FC_OVCAR3" = "OVCAR3"))

heatmap_overlap
```

Saving the heatmap in a png-file in the results folder:

```{r}
ggsave(filename = here("results/05_heatmap_DEGs.png"),
       plot = heatmap_overlap,
       width = 12, 
       height = 8)
```
