==> R/06_analysis_scatterplot.qmd <==
---
title: "06_analysis_scatterplot"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Load packages

```{r}
library(tidyverse)
library(here)
library(readr)
library(scales)
```

### Load data

```{r}
data <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

### Regression models for up- and down-regulated OVCAR3 DEGs

Two separate regression models comparing the OVCAR3 cell line at
baseline and after HERV-K knock-out for both up-regulated DEGs and
down-regulated DEGs.

```{r}
#Regression model for up-regulated DEGs
Equation_OVCAR3_UP <- data |>
  select(OVCAR3_MOCK, 
         OVCAR3_KO, 
         DEG_OVCAR3) |>
  filter(DEG_OVCAR3 == "Upregulated") |>
  lm(formula = OVCAR3_KO ~ OVCAR3_MOCK)

summary(Equation_OVCAR3_UP)


#Regression model for down-regulated DEGs
Equation_OVCAR3_DOWN <- data |>
  select(OVCAR3_MOCK, 
         OVCAR3_KO, 
         DEG_OVCAR3) |>
  filter(DEG_OVCAR3 == "Downregulated") |>
  lm(formula = OVCAR3_KO ~ OVCAR3_MOCK)

summary(Equation_OVCAR3_DOWN)


#Regression model for neutral DEGs
Equation_OVCAR3_NEU <- data |>
  select(OVCAR3_MOCK, 
         OVCAR3_KO, 
         DEG_OVCAR3) |>
  filter(DEG_OVCAR3 == "Neutral") |>
  lm(formula = OVCAR3_KO ~ OVCAR3_MOCK)

summary(Equation_OVCAR3_NEU)
```

### Scatter plot of KO vs baseline DEGs in OVCAR3

A scatter plot is done to visualise the comparison of the OVCAR3 cell
line at baseline and after HERV-K knock-out for both up-regulated DEGs,
down-regulated DEGs and neutral DEGs from above.

```{r}
scatter_plot_OVCAR3 <- data |>  
  ggplot(mapping = aes(x = OVCAR3_MOCK,
                       y = OVCAR3_KO,
                       color = DEG_OVCAR3)) +
  geom_point(size = 1) +
  geom_smooth(method = "lm",
              linewidth = 0.5,
              se = FALSE) +
  labs(title = "Correlation between DEGs at baseline \nand DEGs in HERV-K KO in OVCAR3",
       x = "OVCAR3 baseline", 
       y = "OVCAR3 KO",
       color = "Differentially expressed genes") +
  scale_x_continuous(labels = scientific) +
  scale_y_continuous(labels = scientific) +  
  theme_minimal() + 
  theme(plot.title = element_text(size = 24),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 12),
        legend.key.size = unit(1,"cm")) +
  annotate("text",
           x = 150000, y = 80000, 
           size = 5,
           label = "y = 0.18x - 5.44",
           color = "red") +
  annotate("text",
           x = 150000, y = 40000, 
           size = 5,
           label = "R^2 = 0.95",
           color = "red") +
  annotate("text",
           x = 50000, y = 500000, 
           size = 5,
           label = "y = 5.03x + 13.83",
           color = "blue") + 
  annotate("text",
           x = 50000, y = 460000, 
           size = 5,
           label = "R^2 = 0.99",
           color = "blue") +
  annotate("text",
           x = 150000, y = 400000, 
           size = 5,
           label = "y = 1.74x - 468",
           color = "forestgreen") + 
  annotate("text",
           x = 150000, y = 360000, 
           size = 5,
           label = "R^2 = 0.634",
           color = "forestgreen")

scatter_plot_OVCAR3
```

For all three DEG groups (up- and down-regulated and neutral) a positive
correlation is found between the OVCAR3 cells at baseline and after
HERV-K KO.

The plot is saved in a png-file in the results folder:

```{r}
ggsave(filename = here("results/06_scatter_OVCAR3.png"),
       plot = scatter_plot_OVCAR3,
       width = 8, 
       height = 5)
```

### Regression models for up- and down-regulated SKOV3 DEGs

Two separate regression models comparing the SKOV3 cell line at baseline
and after HERV-K knock-out for both up-regulated DEGs and down-regulated
DEGs.

```{r}
#Regression model for up-regulated DEGs
Equation_SKOV3_UP <- data |>
  select(SKOV3_MOCK, 
         SKOV3_KO, 
         DEG_SKOV3) |>
  filter(DEG_SKOV3 == "Upregulated") |>
  lm(formula = SKOV3_KO ~ SKOV3_MOCK)

summary(Equation_SKOV3_UP)


#Regression model for down-regulated DEGs
Equation_SKOV3_DOWN <- data |>
  select(SKOV3_MOCK, 
         SKOV3_KO, 
         DEG_SKOV3) |>
  filter(DEG_SKOV3 == "Downregulated") |>
  lm(formula = SKOV3_KO ~ SKOV3_MOCK)

summary(Equation_SKOV3_DOWN)


#Regression model for neutral DEGs
Equation_SKOV3_NEU <- data |>
  select(SKOV3_MOCK, 
         SKOV3_KO, 
         DEG_SKOV3) |>
  filter(DEG_SKOV3 == "Neutral") |>
  lm(formula = SKOV3_KO ~ SKOV3_MOCK)

summary(Equation_SKOV3_NEU)
```

### Scatter plot of KO vs baseline DEGs in SKOV3

A scatter plot is done to visualise the comparison of the SKOV3 cell
line at baseline and after HERV-K knock-out for both up-regulated DEGs,
down-regulated DEGs and neutral DEGs from above.

```{r}
#Scatterplot showing all data DEGs
scatter_plot_SKOV3 <- data |> 
  ggplot(mapping = aes(x = SKOV3_MOCK,
                       y = SKOV3_KO,
                       color = DEG_SKOV3)) +
  geom_point(size = 1) +
  geom_smooth(method = "lm",
              se = FALSE,
              linewidth = 0.5) +
  labs(title = "Correlation between DEGs at baseline \nand DEGs HERV-K KO in SKOV3",
       x = "SKOV3 baseline", 
       y = "SKOV3 KO", 
       color = "Differentially\n expressed genes") +
  scale_x_continuous(labels = scientific) +
  scale_y_continuous(labels = scientific) +
  theme_minimal() + 
  theme(plot.title = element_text(size = 24),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 20),
        legend.text = element_text(size = 18))+
  annotate("text",
           x = 55000, y = 110000,
           size = 5,
           label = "y = 5.01x + 4.05",
           color = "blue") +
  annotate("text",
           x = 55000, y = 100000,
           size = 5,
           label = "R^2 = 0.77",
           color = "blue") +
  annotate("text",
           x = 150000, y = 35000,
           size = 5,
           label = "y = 0.20x - 1.08",
           color = "red") +
  annotate("text",
           x = 150000, y = 25000, 
           size = 5,
           label = "R^2 = 0.995",
           color = "red") +
  annotate("text",
           x = 200000, y = 180000, 
           size = 5,
           label = "y = 0.66x + 19.38",
           color = "forestgreen") + 
  annotate("text",
           x = 200000, y = 170000, 
           size = 5,
           label = "R^2 = 0.97",
           color = "forestgreen")

scatter_plot_SKOV3
```

As expected, a clear positive correlation is seen for the neutral DEGs
in SKOV3.

The plot is saved in a png-file in the results folder:

```{r}
ggsave(filename = here("results/06_scatter_SKOV3.png"),
       plot = scatter_plot_SKOV3,
       width = 8, 
       height = 5)
```

### Scatter plot of KO vs baseline DEGs in SKOV3 with changed axis

In the above correlation analysis, it is not possible to see the
down-regulated and up-regulated DEGs in SKOV3, as their relative
expression is low. The same correlation analysis is performed, but with
changed axis making it possible to see the up- and down-regulated DEGs.

```{r}
#Scatterplot with chosen axis FIX: 
scatter_plot_SKOV3_fixed <- data |> 
  ggplot(mapping = aes(x = SKOV3_MOCK,
                       y = SKOV3_KO,
                       color = DEG_SKOV3)) +
  geom_point(size = 1) +
  geom_smooth(method = "lm",
              se = FALSE,
              linewidth = 0.5) +
  theme_minimal() + 
  theme(plot.title = element_text(size = 24),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 20),
        legend.text = element_text(size = 18),
        legend.key = element_rect(color = "white"))+
  labs(title = "Snippet of correlation between DEGs at \nbaseline and DEGs HERV-K KO in SKOV3",
       x = "SKOV3 baseline", 
       y = "SKOV3 KO", 
       color = "Differentially\n expressed genes") +
  scale_x_continuous(labels = scientific) +
  scale_y_continuous(labels = scientific) +  
  xlim(0, 100) +
  ylim(0, 100) +
  annotate("text",
           x = 17, y = 65, 
           size = 5,
           label = "y = 5.01x + 4.05",
           color = "blue") +
  annotate("text",
           x = 17, y = 60, 
           size = 5,
           label = "R^2 = 0.77",
           color = "blue") +
  annotate("text",
           x = 85, y = 23, 
           size = 5,
           label = "y = 0.20x - 1.08",
           color = "red") +
  annotate("text",
           x = 85, y = 18, 
           size = 5,
           label = "R^2 = 0.995",
           color = "red")

scatter_plot_SKOV3_fixed
```

When changing the visible area, we can see similar results as seen for
OVCAR3. A positive correlation is found between the SKOV3 cells at
baseline and after HERV-K KO.

The plot is saved in a png-file in the results folder:

```{r, include=FALSE}
ggsave(filename = here("results/06_scatter_zoom_SKOV3.png"),
       plot = scatter_plot_SKOV3_fixed,
       width = 8, 
       height = 5)
```
