==> R/08_analysis_kegg.qmd <==
---
title: "08_analysis_kegg"
author: "Julie Bruun Brockhoff (s204519), Cecilie Kejlberg Leth (s204584), Amalie Drud Nielsen (s204560), Chrysillis Hellemann Polhaus (s153664), Lise Lotte Eriksen (s154123)"
format: 
  html:
    embed-resources: true
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: inline
---

### Loading packages

```{r}
library(here)
library(readr)
library(tidyverse)
library(biomaRt)
library(clusterProfiler)
library(enrichplot)
```

### Load data

```{r}
data <- read_tsv(file = here("data/03_dat_aug.tsv"))
```

### Get gene IDs for each gene name

Using ensembl and NCBI to retrieve all of the gene IDs for our data. The
NCBI database is chosen because the IDs needs to match with the KEGG
database. The IDs are then joined with the our data based on the gene
names.

```{r}
#connect to ensembl to fetch IDs
mart <- useMart("ensembl", 
                dataset = "hsapiens_gene_ensembl") 

gene_info <- getBM(
  attributes = c("external_gene_name", "entrezgene_id"),
  filters = "external_gene_name",
  values = data |>
    pull(gene),
  mart = mart)

#Join the data and rearrange the new ID column after the old column.
data_with_ID <- data |> 
  right_join(gene_info, 
            by = c("gene" = "external_gene_name")) |>
  relocate(entrezgene_id, 
           .after = gene)

```

### The cell lines are seperated into two different datasets with up- and downregulated DEGs. The DEGs are also ranjed

```{r}
#Separate data into the two cell lines DEGs with geneIDs
data_OVCAR3 <- data_with_ID |> 
  filter(DEG_OVCAR3 == "Upregulated" |
           DEG_OVCAR3 == "Downregulated") |>
  dplyr::select(gene, entrezgene_id, OVCAR3_MOCK, OVCAR3_KO, log2FC_OVCAR3) |> 
  mutate(rank_log2FC_OVCAR3 = rank(-log2FC_OVCAR3, ties.method = "min")) |> 
   arrange(rank_log2FC_OVCAR3)

data_SKOV3 <- data_with_ID |> 
  filter(DEG_SKOV3 == "Upregulated" |
           DEG_SKOV3 == "Downregulated") |> 
  dplyr::select(gene, entrezgene_id, SKOV3_MOCK, OVCAR3_KO, log2FC_SKOV3) |>
  mutate(rank_log2FC_SKOV3 = rank(-log2FC_SKOV3, ties.method = "min")) |> 
  arrange(rank_log2FC_SKOV3)


#save the geneIDs seperately for each cell line and type of regulation:
geneIDs_OVCAR3 <- data_OVCAR3 |> 
  rename(geneIDs = entrezgene_id) |>  
  pull(geneIDs) |> 
  na.omit()

geneIDs_SKOV3 <- data_SKOV3 |> 
  rename(geneIDs = entrezgene_id) |>  
  pull(geneIDs) |> 
  na.omit()

```

### KEGG enrichment for OVCAR3

```{r}
kegg_analysis_OVCAR3 <- enrichKEGG(gene = geneIDs_OVCAR3,
                                               organism = "hsa",
                                               pAdjustMethod = "BH",
                                               minGSSize = 10,
                                               maxGSSize = 500,
                                               qvalueCutoff = 0.05) |> 
  as_tibble()

plot_kegg_OVCAR3 <- kegg_analysis_OVCAR3 |> 
  slice_min(order_by = p.adjust, 
            n = 10) |> 
  mutate(Description = fct_reorder(Description, p.adjust, .desc = TRUE)) |>
  ggplot(aes(x = Count, 
             y = Description)) + 
  geom_point(aes(size = Count, 
                 color = p.adjust)) + 
  scale_color_gradient(low = "red", 
                       high = "blue", 
                       name = "p.adjust") + 
  labs(
    title = "KEGG Pathway Enrichment Analysis for \n genes in OVCAR3",
    x = "Gene Count",
    y = "Pathways") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 24,
                              hjust = 0.8),
    axis.text.x = element_text(size = 18,
                             angle = 45),
    axis.text.y = element_text(size = 18),
    axis.title = element_text(size = 18),
    legend.title = element_text(size = 16),
    legend.text = element_text(size = 14))

plot_kegg_OVCAR3

```

Several pathways was significantly for the OVCAR3 cell line for the up-
and down-regulated genes. The most prominent pathways are related the
cancer along with PI3K-Akt signaling pathway, this pathway is a pathway
that regulates cell growth, apoptosis and cellular senescence.

The KEGG is saved in a png-file in the results folder:

```{r}
ggsave(filename = here("results/08_KEGG_OVCAR3.png"),
       plot = plot_kegg_OVCAR3,
       width = 10,
       height = 6)
```

### KEGG enrichment for SKOV3

```{r}
kegg_analysis_SKOV3 <- enrichKEGG(gene = geneIDs_SKOV3,
                                               organism = "hsa",
                                               pAdjustMethod = "BH",
                                               minGSSize = 10,
                                               maxGSSize = 500,
                                               qvalueCutoff = 0.05) |> 
  as_tibble()

plot_kegg_SKOV3 <- kegg_analysis_SKOV3 |> 
  slice_min(order_by = p.adjust, 
            n = 10) |> 
  mutate(Description = fct_reorder(Description, p.adjust, .desc = TRUE)) |>
  ggplot(aes(x = Count, 
             y = Description)) + 
  geom_point(aes(size = Count, 
                 color = p.adjust)) + 
  scale_color_gradient(low = "red", 
                       high = "blue", 
                       name = "p.adjust") + 
  labs(
    title = "KEGG Pathway Enrichment Analysis for \n genes in SKOV3",
    x = "Gene Count",
    y = "Pathways") +
  theme_minimal() +
  theme(
    plot.title = element_text(size = 24,
                              hjust = 0.8),
    axis.text.x = element_text(size = 18,
                             angle = 45),
    axis.text.y = element_text(size = 18),
    axis.title = element_text(size = 18),
    legend.title = element_text(size = 16),
    legend.text = element_text(size = 14))

plot_kegg_SKOV3

```

There are no pathways found for the SKOV3 cell line.

The KEGG is saved in a png-file in the results folder:

```{r}
ggsave(filename = here("results/08_KEGG_SKOV3.png"),
       plot = plot_kegg_SKOV3,        
       width = 10,        
       height = 6)
```

### Conclusion

Enrichment analysis was done for the two cell lines, OVCAR3 and SKOV3.
The DEGs for each cell line were ranked based on their log2FC and a KEGG
analysis was done for both cell lines. The results from the OVCAR3 cell
line showed several pathway, mostly related to cancer along with the
PI3K-Akt signaling pathway. This pathway is usually down-regulated
during cellular senescence. This could indicate that the KO OVCAR3 cell
line doesn't promote cancer and cellular senescence as well as the
baseline cell line (MOCK).

The enrichment analysis of SKOV3 didn't reveal any significant pathways.

These results indicate that the KO OVCAR3 cell line and the baseline
OVCAR3 cell is a better model for cancer compared to SKOV3 as this cell
line doesn't show any significant differences between the two set of
DEGs in the baseline compared to the KO cell line.
